pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tinydiffusion"
version = "0.1.0"
description = "A lightweight diffusion model for generating MNIST digit images"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "TinyDiffusion Team"}
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "torch>=2.5.1",
    "torchvision>=0.20.1",
    "torchaudio>=2.5.1",
    "numpy>=2.2.4",
    "matplotlib>=3.10.1",
    "pillow>=11.1.0",
    "tqdm>=4.67.1",
    "wandb>=0.19.8",
    "scipy>=1.15.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
]

[tool.setuptools]
packages = ["model", "src", "utils"]

[tool.black]
line-length = 88
target-version = ["py38"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
